chore: move conv2d layer into layers file, added layer to mnist example
